MTFBASE 11-04-23 11:40

+++ database/migrations/2023_03_09_161228_create_transaction_masters_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->double('amount_commission');  //MONTO COMISION
             $table->double('exchange_rate')->nullable(); // TAZA DE CAMBIO
             $table->enum('exonerate', [1, 2, 3])->nullable()->default(1); //DESCUENTO, EXONERADO E INCLUIR COMOSIÓN
-            $table->integer('percentage'); //PORCENTAJE DE LA TRANSFERENCIA
+            $table->double('percentage'); //PORCENTAJE DE LA TRANSFERENCIA
             $table->foreignId('type_coin_id')->references('id')->on('type_coins')->nullable(); // TIPO DE MONEDA DE LA TRANSFERENCIA
             $table->foreignId('type_transaction_id')->references('id')->on('type_transactions')->nullable();  //TIPO DE LA TRANSFERENCIA
             $table->foreignId('user_id')->references('id')->on('users')->nullable();  // USUARIO QUE REALIZO LA TRANSFERENCIA
@@ -28,7 +28,7 @@ return new class extends Migration
             $table->enum('status', ['Activo', 'Anulado'])->nullable()->default('Activo'); //ESTATUS
             $table->string('description'); //DESCRIPCION DE LA TRANSFERENCIA
             $table->date('transaction_date'); //FECHA DE LA TRANSACCIÓN
-            $table->integer('percentage_base')->nullable(); //PORCENTAJE DE LA GANANCIA
+            $table->double('percentage_base')->nullable(); //PORCENTAJE DE LA GANANCIA
             $table->double('amount_commission_base')->nullable(); //GANANCIA DE LA COMOSION
             $table->timestamps(); //CREACION Y MODIFICACION POR DEFECTO (TOMADO POR EL SERVIDOR)
         });
